requirements.txt
openai
python-dotenv
fastapi
uvicorn 